Rename project llvm.nativeint
The new project is called llvm.context and also contains context classes like LLVMContext or LLVMFunctionRegistry.
