+++ projects/com.oracle.truffle.llvm.context/src/com/oracle/truffle/llvm/context/LLVMLanguage.java
@@ -27,7 +27,7 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
  * OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.oracle.truffle.llvm.nodes.impl.base;
+package com.oracle.truffle.llvm.context;
 
 import java.io.IOException;
