+++ projects/com.oracle.truffle.llvm.context/src/com/oracle/truffle/llvm/context/nativeint/NativeLookup.java
@@ -27,7 +27,7 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
  * OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.oracle.truffle.llvm.nativeint;
+package com.oracle.truffle.llvm.context.nativeint;
 
 import java.lang.reflect.Method;
 import java.util.ArrayList;
