+++ projects/com.oracle.truffle.llvm.test/src/com/oracle/truffle/llvm/test/TestGCCCompileSuite.java
@@ -40,7 +40,7 @@ import org.junit.runners.Parameterized;
 import com.oracle.truffle.api.source.Source;
 import com.oracle.truffle.api.vm.PolyglotEngine;
 import com.oracle.truffle.api.vm.PolyglotEngine.Builder;
-import com.oracle.truffle.llvm.nodes.impl.base.LLVMLanguage;
+import com.oracle.truffle.llvm.context.LLVMLanguage;
 import com.oracle.truffle.llvm.runtime.LLVMLogger;
 
 @RunWith(Parameterized.class)
