+++ projects/com.oracle.truffle.llvm/src/com/oracle/truffle/llvm/SulongLibrary.java
@@ -39,7 +39,7 @@ import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
 
 import com.oracle.truffle.api.source.Source;
-import com.oracle.truffle.llvm.nodes.impl.base.LLVMLanguage;
+import com.oracle.truffle.llvm.context.LLVMLanguage;
 
 public class SulongLibrary {
